# apps/onboarding-bff/src/main/openapi/onboarding.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Onboarding API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/v1/onboarding" : {
      "get" : {
        "tags" : [ "Onboarding Controller", "external-v2" ],
        "summary" : "Retrieve paged onboardings with optional filters and sorting.",
        "description" : "The API retrieves paged onboarding using optional filter, order by descending creation date",
        "operationId" : "getOnboardingWithFilter",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "institutionId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "onboardingId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "0",
            "type" : "integer"
          }
        }, {
          "name" : "productId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "productIds",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "20",
            "type" : "integer"
          }
        }, {
          "name" : "skipPagination",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OnboardingStatus"
          }
        }, {
          "name" : "subunitCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taxCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingGetResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Onboarding Controller" ],
        "summary" : "Default onboarding for GSP/SA/AS institutions, saves user data, and triggers async onboarding activities.",
        "description" : "Perform default onboarding request, it is used for GSP/SA/AS institution type.Users data will be saved on personal data vault if it doesn't already exist.At the end, function triggers async activities related to onboarding based on institution type.",
        "operationId" : "onboarding",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingDefaultRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AdditionalInformationsDto" : {
        "type" : "object",
        "properties" : {
          "belongRegulatedMarket" : {
            "type" : "boolean"
          },
          "regulatedMarketNote" : {
            "type" : "string"
          },
          "ipa" : {
            "type" : "boolean"
          },
          "ipaCode" : {
            "type" : "string"
          },
          "establishedByRegulatoryProvision" : {
            "type" : "boolean"
          },
          "establishedByRegulatoryProvisionNote" : {
            "type" : "string"
          },
          "agentOfPublicService" : {
            "type" : "boolean"
          },
          "agentOfPublicServiceNote" : {
            "type" : "string"
          },
          "otherNote" : {
            "type" : "string"
          }
        }
      },
...
      "AggregateInstitutionRequest" : {
        "required" : [ "taxCode", "description" ],
        "type" : "object",
        "properties" : {
          "taxCode" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "subunitCode" : {
            "type" : "string"
          },
          "subunitType" : {
            "type" : "string"
          },
          "vatNumber" : {
            "type" : "string"
          },
          "parentDescription" : {
            "type" : "string"
          },
          "geographicTaxonomies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GeographicTaxonomy"
            }
          },
          "address" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          },
          "originId" : {
            "type" : "string"
          },
          "origin" : {
            "$ref" : "#/components/schemas/Origin"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "istatCode" : {
            "type" : "string"
          },
          "recipientCode" : {
            "type" : "string"
          },
          "digitalAddress" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "county" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "isTest" : {
            "type" : "boolean"
          }
        }
      },
...
      "BillingRequest" : {
        "type" : "object",
        "properties" : {
          "vatNumber" : {
            "type" : "string"
          },
          "recipientCode" : {
            "type" : "string"
          },
          "publicServices" : {
            "type" : "boolean"
          }
        }
      },
      "BillingResponse" : {
        "type" : "object",
        "properties" : {
          "vatNumber" : {
            "type" : "string"
          },
          "recipientCode" : {
            "type" : "string"
          },
          "publicServices" : {
            "type" : "boolean"
          }
        }
      },
...
      "DataProtectionOfficerRequest" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "pec" : {
            "type" : "string"
          }
        }
      },
      "GPUData" : {
        "type" : "object",
        "properties" : {
          "businessRegisterNumber" : {
            "type" : "string"
          },
          "legalRegisterNumber" : {
            "type" : "string"
          },
          "legalRegisterName" : {
            "type" : "string"
          },
          "longTermPayments" : {
            "type" : "boolean"
          },
          "manager" : {
            "type" : "boolean"
          },
          "managerAuthorized" : {
            "type" : "boolean"
          },
          "managerEligible" : {
            "type" : "boolean"
          },
          "managerProsecution" : {
            "type" : "boolean"
          },
          "institutionCourtMeasures" : {
            "type" : "boolean"
          }
        }
      },
      "GeographicTaxonomy" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "desc" : {
            "type" : "string"
          }
        }
      },
      "GeographicTaxonomyDto" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "desc" : {
            "type" : "string"
          }
        }
      },
...
      "InstitutionBaseRequest" : {
        "required" : [ "institutionType", "origin", "originId", "digitalAddress" ],
        "type" : "object",
        "properties" : {
          "institutionType" : {
            "$ref" : "#/components/schemas/InstitutionType"
          },
          "taxCode" : {
            "type" : "string"
          },
          "subunitCode" : {
            "type" : "string"
          },
          "subunitType" : {
            "$ref" : "#/components/schemas/InstitutionPaSubunitType"
          },
          "origin" : {
            "$ref" : "#/components/schemas/Origin"
          },
          "originId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "county" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "digitalAddress" : {
            "minLength" : 1,
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          },
          "istatCode" : {
            "type" : "string"
          },
          "legalForm" : {
            "type" : "string"
          },
          "geographicTaxonomies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GeographicTaxonomyDto"
            }
          },
          "rea" : {
            "type" : "string"
          },
          "shareCapital" : {
            "type" : "string"
          },
          "businessRegisterPlace" : {
            "type" : "string"
          },
          "supportEmail" : {
            "type" : "string"
          },
          "supportPhone" : {
            "type" : "string"
          },
          "imported" : {
            "type" : "boolean"
          },
          "atecoCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "isTest" : {
            "type" : "boolean"
          }
        }
      },
...
      "InstitutionPaSubunitType" : {
        "enum" : [ "AOO", "UO", "EC" ],
        "type" : "string"
      },
...
      "InstitutionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "institutionType" : {
            "type" : "string"
          },
          "taxCode" : {
            "type" : "string"
          },
          "taxCodeInvoicing" : {
            "type" : "string"
          },
          "subunitCode" : {
            "type" : "string"
          },
          "subunitType" : {
            "$ref" : "#/components/schemas/InstitutionPaSubunitType"
          },
          "origin" : {
            "$ref" : "#/components/schemas/Origin"
          },
          "originId" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "county" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "digitalAddress" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          },
          "parentDescription" : {
            "type" : "string"
          },
          "geographicTaxonomies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GeographicTaxonomyDto"
            }
          },
          "rea" : {
            "type" : "string"
          },
          "shareCapital" : {
            "type" : "string"
          },
          "businessRegisterPlace" : {
            "type" : "string"
          },
          "supportEmail" : {
            "type" : "string"
          },
          "supportPhone" : {
            "type" : "string"
          },
          "paymentServiceProvider" : {
            "$ref" : "#/components/schemas/PaymentServiceProviderRequest"
          },
          "dataProtectionOfficer" : {
            "$ref" : "#/components/schemas/DataProtectionOfficerRequest"
          },
          "atecoCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "legalForm" : {
            "type" : "string"
          }
        }
      },
      "InstitutionType" : {
        "enum" : [ "PA", "PG", "GSP", "SA", "PT", "SCP", "PSP", "AS", "REC", "CON", "PRV", "PRV_PF", "GPU", "SCEC" ],
        "type" : "string"
      },
      "LocalDateTime" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T12:15:50"
      },
...
      "OnboardingDefaultRequest" : {
        "required" : [ "productId", "institution" ],
        "type" : "object",
        "properties" : {
          "productId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "aggregates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AggregateInstitutionRequest"
            }
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "userRequester" : {
            "$ref" : "#/components/schemas/UserRequesterDto"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionBaseRequest"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingRequest"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "gpuData" : {
            "$ref" : "#/components/schemas/GPUData"
          }
        }
      },
      "OnboardingGet" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionResponse"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingResponse"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "expiringDate" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "activatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "status" : {
            "type" : "string"
          },
          "userRequester" : {
            "$ref" : "#/components/schemas/UserRequesterResponse"
          },
          "reasonForReject" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "OnboardingGetResponse" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OnboardingGet"
            }
          }
        }
      },
...
      "OnboardingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionResponse"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserOnboardingResponse"
            }
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingResponse"
          },
          "status" : {
            "type" : "string"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "referenceOnboardingId" : {
            "type" : "string"
          },
          "userRequester" : {
            "$ref" : "#/components/schemas/UserRequesterResponse"
          }
        }
      },
      "OnboardingStatus" : {
        "enum" : [ "REQUESTING", "REQUEST", "TOBEVALIDATED", "PENDING", "PENDING_IN_REVIEW", "COMPLETED", "FAILED", "REJECTED", "DELETED", "OVERRIDDEN" ],
        "type" : "string"
      },
...
      "Origin" : {
        "enum" : [ "MOCK", "IPA", "SELC", "ANAC", "UNKNOWN", "ADE", "INFOCAMERE", "IVASS", "PDND_INFOCAMERE" ],
        "type" : "string"
      },
      "PartyRole" : {
        "enum" : [ "MANAGER", "DELEGATE", "SUB_DELEGATE", "OPERATOR", "ADMIN_EA" ],
        "type" : "string"
      },
      "PaymentServiceProviderRequest" : {
        "type" : "object",
        "properties" : {
          "businessRegisterNumber" : {
            "type" : "string"
          },
          "legalRegisterNumber" : {
            "type" : "string"
          },
          "legalRegisterName" : {
            "type" : "string"
          },
          "longTermPayments" : {
            "type" : "boolean"
          },
          "abiCode" : {
            "type" : "string"
          },
          "vatNumberGroup" : {
            "type" : "boolean"
          },
          "providerNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "contractType" : {
            "type" : "string"
          },
          "contractId" : {
            "type" : "string"
          }
        }
      },
...
      "UserOnboardingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/PartyRole"
          },
          "productRole" : {
            "type" : "string"
          },
          "userMailUuid" : {
            "type" : "string"
          }
        }
      },
      "UserRequest" : {
        "type" : "object",
        "properties" : {
          "taxCode" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/PartyRole"
          }
        }
      },
      "UserRequesterDto" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
      "UserRequesterResponse" : {
        "type" : "object",
        "properties" : {
          "userMailUuid" : {
            "type" : "string"
          },
          "userRequestUid" : {
            "type" : "string"
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "taxCode" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/PartyRole"
          },
          "productRole" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}